murriel/codelabs/requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
google-api-core==1.17.0
google-auth==1.16.0
google-cloud-vision==1.0.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
idna==2.9
numpy==1.18.5
pgzero==1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.0.0.dev4
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.15.0
urllib3==1.25.9
